chat_bot/movie_search/requirements.txt
streamlit
python-dotenv
google-generativeai
python-dotenv
langchain
langchain_google_genai